+++ core/datastore/src/main/java/com/ohnalmwo/datastore/datasource/location/LocationsDataSource.kt
@@ -1,4 +1,4 @@
-package com.ohnalmwo.datastore.datasource
+package com.ohnalmwo.datastore.datasource.location
 
 import com.ohnalmwo.model.LocationInfo
 import kotlinx.coroutines.flow.Flow
